requirements.txt
gensim
matplotlib
numpy
pandas
rdflib
scikit_learn
python-louvain
rdflib==5.0.0
pickle
json
argparse
tensorflow-gpu==1.15.0
sentencepiece==0.1.85
tqdm==4.41.0
networkx==2.4
loguru==0.4.1

